+++ packages/riverpod/lib/riverpod.dart
@@ -6,6 +6,10 @@ export 'src/common.dart'
         AsyncData,
         AsyncLoading,
         AsyncError,
+        Result,
+        ResultX,
+        ResultData,
+        ResultError,
         AsyncValueX,
         AsyncValueLoadingError;